+++ plugin.image.mypicsdb3/resources/lib/mypicsdb3/__init__.py
@@ -1,5 +1,5 @@
 """Core package for MyPicsDB 3."""
 
 ADDON_ID = "plugin.image.mypicsdb3"
-VERSION = "0.8.21"
+VERSION = "0.8.22"
 SCHEMA_VERSION = 9